// frontend/src/shared/lib/api/model/runMapSummaryResponse.ts
/**
 * Generated by orval. Do not edit manually.
 * daikon-studio
 */
import type { RunMapSummaryResponseNearestMin } from './runMapSummaryResponseNearestMin';
import type { RunMapSummaryResponseNearestMax } from './runMapSummaryResponseNearestMax';

export interface RunMapSummaryResponse {
  total: number;
  in_domain: number;
  threshold: number;
  nearest_min: RunMapSummaryResponseNearestMin;
  nearest_max: RunMapSummaryResponseNearestMax;
}